requirements.txt
absl-py==2.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
grpcio==1.76.0
gym==0.26.2
gym-notices==0.1.0
gymnasium==1.2.2
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
Markdown==3.10
MarkupSafe==2.1.5
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
nptyping==2.5.0
numpy==1.26.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==25.0
pillow==12.0.0
protobuf==6.33.1
pydot==4.0.1
pygame==2.6.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scikit-learn==1.7.2
scipy==1.16.3
setuptools==70.2.0
shapely==2.1.2
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
torch==2.9.1+cu126
torchvision==0.24.1+cu126
tqdm==4.67.1
triton==3.5.1
typing_extensions==4.15.0
Werkzeug==3.1.3
